(maint) Fix fact output error in environment acceptance tests
Prior to this commit we were experiencing a failure in the
environment acceptance testing due to a recent change that reduced
fact loading output to debug only.

Now the test is no longer failing since we are now explicitly
loading the global module instead of using the fact loading output.

Paired with Hailee Kenney <hailee@puppetlabs.com>