(PUP-2971) Fix yum package provider missing install_options for list.
When executing the list command to filter virtual packages (when
allow_virtual is false or nil), the yum package provider was not passing
down the install options to the list command.  For users passing options
like --enablerepo, this prevents the packages from being discovered and
the list command fails.

The fix is to properly pass down the install options when executing the
list command.