(PUP-2040) .pmtignore and .gitignore for build
Prior to this commit there was no way for a module developer to specify files
that they want to ignore when building/packaging their module.

This commit adds the ability to specify files to ignore in a .pmtignore file,
located in the project root and using the same syntax as .gitignore, or if that
is not present, reading the .gitignore file and using that.

This commit also adds pathspec as a vendored gem. (PE-5574)