(PUP-2907) Ignore source permissions on windows for fact sync
Previously, puppet on windows would always issue a warning:

    Copying owner/mode/group from the source file on Windows is
    deprecated; use source_permissions => ignore

Regardless of what was in the user's catalog. The reason is because
puppet applies an internal catalog to download external facts, and that
catalog contains a file resource whose source_permissions parameter is
set to `:use`.

This is done because external facts are often times scripts, which must
be executable for puppet to invoke them. See commit 21edcd67b for
PUP-2705.

This commit makes it so that source permissions for external facts are
ignored on Windows, while preserving existing behavior on other
platforms.

Permissions can be ignored on Windows, because puppet always has full
control of files it creates.