(PUP-3036) Rubocop Style/AndOr autocorrect pass
This commit replaces `and` and `or` keywords in favor of `&&` and `||` where it
can be done automatically (In places where the semantic meaning of the
expression does not change on replacing the `and` keyword with the operator `||`).

Since we are using `rubocop -a` to make the changes, we are assured of the
correctness of translation. These changes have also been individually verified.

We have been able to get rubocop to make further changes than previously by
identifying equivalent patterns identified as different in parsing
(See rubocop issue #1255)