(PUP-3036) Rubocop Style/Not automatic pass
This commit replaces `not` in favor of `!` where
can be done automatically (In places where the semantic meaning of the
expression does not change on replacing the `not` keyword with the operator `!`).

Since we are using `rubocop -a` to make the changes, we are assured of the
correctness of translation. These changes have also been individually verified.

This change is especially useful because after replacing such instances, the
AndOr cop can autocorrect even more instances of violations

This also enables the Style/Not cop