(PUP-3036) Rubocop Style/AndOr automatic pass (2)
This commit replaces `and` and `or` keywords in favor of `&&` and `||` where it
can be done automatically (In places where the semantic meaning of the
expression does not change on replacing the `and` keyword with the operator `||`).

Since we are using `rubocop -a` to make the changes, we are assured of the
correctness of translation. These changes have also been individually verified.

This commit takes advantage of the previous changes replacing `not` keyword to
autocorrect even more instances of violations.