(PUP-3036) Rubocop Style/AndOr parenthesis automatic pass
This commit replaces `AndOr` in favor of `and` and `or` where
can be done automatically (In places where the semantic meaning of the
expression does not change on replacing the `and` keyword with the operator `or`).

Since we are using `rubocop -a` to make the changes, we are assured of the
correctness of translation. These changes have also been individually verified.

The difference in this commit is that, we specifically search for instances
where the semantics of `and` or `or` is unchanged from `&&` and `||` when
parenthesis are introduced for method calls. That is, we look for instances
of the form `mymethod a,b or mymethod c` and convert them to
`mymethod( a,b ) or mymethod( c)` before applying the transformation.