(PUP-2560) Disallow facts over the network
The final removal of the inventory service is to stop allowing facts
over the network. This stops the various facts terminii from responding.
Technically it was not longer reachable anyway because of the
pluralization rule changes: facts is seen as a "search" query for the "fact"
indirection, which does not exist. However, just to be sure, this stops
them from being used at all, even if there was a request constructed
that could get the facts indirection.