(PUP-3063) Remove client certificate expiration check
This is a follow-up to PUP-3031.  The code that we were using
to issue a warning about expiration dates of client certificates
was in a pretty heavily-frequented path, was expensive, and
did not account for differences between the various web
servers that we might be running in.  We decided that it would
be better to provide a CLI tool for this, rather than having
it be something that is triggered on each HTTP request.

The design / scope for the CLI tool is not fleshed out yet,
and we did not want to leave orphaned code in the source tree
until that happens, so the code is just being removed for now.