(PUP-3030) Rename P::F::Tempfile to Uniquefile
Since the new implementation of Puppet::FileSystem::Tempfile is
not trying to manage the deletion of files created through it,
the name was misleading.  This commit simply changes the name
to Uniquefile to try to make it a bit more obvious that it is
not responsible for deleting the files.