(maint) Use tlsv1 when curling the master from another agent
Various acceptance tests curl the puppet master, and rely on the
installed openssl lib.  With the inclusion of PUP-417, we began
installing the puppetmaster-passenger packages on debian, which
ultimately rely on updated openssl packages being installed on the
puppet master node.  If the agent platfrom was old enough to be using
openssl 0.9.8 (lucid for example), several curl tests began failing
because they were attempting to connect with sslv2 and being rejected.

This patch updates the five failing tests to be explicit about requiring
--tlsv1.