(PUP-2914) Remove remaining rgen dependencies
In 6a94700, the rgen 0.7.0 library was added to the vendor directory in puppet,
and as such is no longer an external dependency. This commit removes the rgen
package as a dependency of the puppet rpm and deb packages, so that it is not
installed in addition to the vendored rgen (or worse even installed at a
conflicting version).

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>