(PUP-3127) LDAP - introduce LDAP-Certificat Directory
Introduction of a new config parameter 'ldapcrtdir' to implement ssl based ldap connections on Solaris based Puppet-Master.
Without this implementation it is not possible to create ssl encrypted ldap connections on a Soalris system.
If the 'ldapcrtdir' parameter is not set or set to false (=> default), the connectin is set up as before.