(PUP-2349) Force metadata mode to be an octal string
The initial commit for this ticket missed the case where
the mode value isn't explicitly specified but rather comes from
source permissions of a file resource, e.g. as happens during
pluginsync.

This change ensures that Puppet::FileServing::Metadata#mode is
converted to an octal string if it's a Numeric.