(maint) Clear record of configured deprecated settings
We cache a list of deprecated settings that have been configured in the
initial parse_config, so that later initialization steps allow us to
read the :disable_warnings value, and then decide whether or not to
actually issue a particular warning.

But we weren't clearing this cache in the unsafe_clear call, which was
allowing this state to leak across specs.  Once we added that cache
clear to unsafe_clear, we then had to reorder parse_config steps for
unsafe_clear, followed by record_deprecations_from_puppet_conf(data).