(PUP-3069) Fix path expansion for environment.conf test on Win
The intention in these specs is for the default_manifest setting
to be absolute for the platform under test.  Without an expand_path,
the POSIX style setting we're using will still seem relative on
windows, causing these two tests to fail.  Really these settings
should only be used on a master instance, which would be POSIX,
but for completeness sake, this commit expands the initial
default_manifest paths.