Optimize the read of the system mounts in selinux
Util::SELinux currently reads /proc/mounts for each file in the system.

This can cause a massive lag/burden on the system especially if the
system uses a great number of mount entries (such as OpenShift).

However, this is not a great patch since it creates a global variable
insted of using whatever global caching mechanism that Puppet may have
internally.

Submitting for review on PUP-3116 but hoping for improvement.