(PUP-987) Remove the Ruby DSL
This removes the deprecated Ruby DSL support. It also removes
deprecation checks; an attempt to use a .rb manifest somewhere will
simply fail in some undefined manner, just as if there was an attempt to
use some other arbitrary text file. Most likely a syntax error from the
parser if someone forced it to use a .rb file.