(PUP-3136) Fix Intermittent Acceptance Test Failures
It was discovered that when puppet attempted to start apache the
provider would say that it was stopped even though it was starting. On
agents where a puppet master was installed with passenger the apache
service would take longer than usual to start lengthening the starting
process. By the time puppet tested if the service was running or not
apache would still be attempting to start causing the test to fail.

This commit replaces the apache2 package with rabbitmq since it is not
used by puppet and therefore tests down the road will not fail with us
managing it with puppet. It also uses a manifest instead of the
`puppet_resource` function to ensure that there is not any more timing
problems giving us the errors.