(PUP-900) Don't process refresh events for noop'ed resources
When resources were processing queued events, the event callbacks would
be invoked if any of the events were not noop regardless of the state of
the resource in question. This means that a resource that had `noop =>
true` would have the refresh action triggered on refresh, bypassing the
noop statement.

This commit changes the refresh behavior to ensure that resources that
are set to noop are not refreshed.